BETA: Ugly temp fix for augment skills disappearing from the shortcuts.
git-svn-id: http://svn.l2jserver.com/branches/unstable@5129 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/model/actor/instance/L2PcInstance.java
@@ -8124,7 +8124,8 @@ public final class L2PcInstance extends L2Playable
 		final L2ShortCut[] allShortCuts = getAllShortCuts();
 		for (L2ShortCut sc : allShortCuts)
 		{
-			if ((sc != null) && (skill != null) && (sc.getId() == skill.getId()) && (sc.getType() == L2ShortCut.TYPE_SKILL))
+			if ((sc != null) && (skill != null) && (sc.getId() == skill.getId()) && (sc.getType() == L2ShortCut.TYPE_SKILL)
+				&& !(skill.getId() >= 3080 && skill.getId() <= 3259)) // FIXME: Rough fix for shortcuts of augments getting removed. Find a better way
 				deleteShortCut(sc.getSlot(), sc.getPage());
 		}
 		return oldSkill;
